+++ src/aux-vm-browser/managers/BotManager.ts
@@ -26,6 +26,7 @@ import {
     ConnectionIndicator,
     getConnectionId,
     DEFAULT_BRANCH_NAME,
+    BotsState,
 } from '@casual-simulation/aux-common';
 import {
     AuxVM,
@@ -38,7 +39,6 @@ import {
 } from '@casual-simulation/aux-vm';
 import { BotPanelManager } from './BotPanelManager';
 import { BrowserSimulation } from './BrowserSimulation';
-import { AuxVMImpl } from '../vm/AuxVMImpl';
 import { PortalManager, ProgressManager } from '@casual-simulation/aux-vm';
 import { filter, tap, map } from 'rxjs/operators';
 import { ConsoleMessages } from '@casual-simulation/aux-common';
@@ -136,7 +136,8 @@ export class BotManager extends BaseSimulation implements BrowserSimulation {
         id: string,
         configBotId: string,
         origin: SimulationOrigin,
-        config: AuxConfig['config']
+        config: AuxConfig['config'],
+        initialTempState: BotsState = {}
     ): Partial<AuxPartitionConfig> {
         const defaultPartitions: Partial<AuxPartitionConfig> = {
             [TEMPORARY_BOT_PARTITION_ID]: {
@@ -146,6 +147,7 @@ export class BotManager extends BaseSimulation implements BrowserSimulation {
                     [configBotId]: createBot(configBotId, {
                         inst: origin.inst ?? id,
                     }),
+                    ...initialTempState,
                 },
             },
             [COOKIE_BOT_PARTITION_ID]: {
@@ -243,7 +245,8 @@ export class BotManager extends BaseSimulation implements BrowserSimulation {
         id: string,
         configBotId: string,
         origin: SimulationOrigin,
-        config: AuxConfig['config']
+        config: AuxConfig['config'],
+        initialTempState: BotsState = {}
     ): AuxPartitionConfig {
         const localPersistence = config.staticRepoLocalPersistence ?? true;
         console.log('[BotManager] Using static partitions');
@@ -256,7 +259,8 @@ export class BotManager extends BaseSimulation implements BrowserSimulation {
             id,
             configBotId,
             origin,
-            config
+            config,
+            initialTempState
         );
 
         let partitions: AuxPartitionConfig = {